update : corrected limits adder test
