+++ iidm/iidm-tck/src/test/java/com/powsybl/iidm/network/tck/AbstractCurrentLimitsTest.java
@@ -739,7 +739,7 @@ public abstract class AbstractCurrentLimitsTest extends AbstractIdenticalLimitsT
         adder.add();
 
         assertTrue(areLimitsIdentical(limits1, limits2));
-        assertThrows(PowsyblException.class, () -> line.newCurrentLimits1(null));
+        assertFalse(line.newCurrentLimits1(null).hasTemporaryLimits());
     }
 
     @Test
